refactor file utils and fix open file creates subprocess in windows
