+++ shared/utils/src/main/kotlin/ir/amirab/util/osfileutil/LinuxFileUtils.kt
@@ -1,0 +1,31 @@
+package ir.amirab.util.osfileutil
+
+import java.io.File
+
+internal class LinuxFileUtils : FileUtilsBase() {
+    override fun openFileInternal(file: File): Boolean {
+        return execAndWait(arrayOf("xdg-open", file.path))
+    }
+
+    override fun openFolderOfFileInternal(file: File): Boolean {
+        val dbusSendResult = execAndWait(
+            arrayOf(
+                "dbus-send",
+                "--print-reply",
+                "--dest=org.freedesktop.FileManager1",
+                "/org/freedesktop/FileManager1",
+                "org.freedesktop.FileManager1.ShowItems",
+                "array:string:file://${file.path}",
+                "string:"
+            )
+        )
+        if (dbusSendResult) {
+            return true
+        }
+        val xdgOpenResult = execAndWait(
+            arrayOf("xdg-open", file.parent)
+        )
+        return xdgOpenResult
+    }
+
+}
